refactor : 패키지명 변경
요구사항에 맞춰서 변경
