+++ src/main/kotlin/racing/view/UserInput.kt
@@ -1,3 +1,3 @@
-package racing.ui
+package racing.view
 
 data class UserInput(val carNames: List<String>, val moveCount: Int)
